requirements.txt
neat-python==0.92
numpy==1.22.0
pathfinding==1.0.1
pygame==2.1.2
